ARIA-main/ARIA website/requirements.txt
Flask==3.1.0
requests==2.32.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
